+++ tests/testthat/test-zzz.R
@@ -27,8 +27,20 @@ test_that("can construct from unions", {
   expect_equal(new_union(u1, integer), u3)
 })
 
+test_that("base unions print as expected", {
+  expect_snapshot({
+    base_unions
+  })
+})
+
 test_that("can construct from S3 and S4 classes", {
   s4_union <- methods::setClass("s4_union")
   u <- new_union(s3_factor, s4_union)
   expect_equal(u@classes, list(s3_factor, getClass("s4_union")))
+})
+
+test_that("base classes types check their data", {
+  expect_snapshot(error = TRUE, {
+    base_classes$integer(TRUE)
+  })
 })