+++ NEWS.md
@@ -2,9 +2,15 @@
 
 ## Jan 2020
 
+* New `props<-` for setting multiple properties simultaneously and validating
+  afterwards (#149).
+* Validation now happens recursively, and validates types before validating 
+  the object (#149)
 * Classes (base types, S3, S4, and R7) are handled consistently wherever they
   are used. Strings now only refer to base types. New explicit `s3_class()` for 
-  referring to S3 classes (#134).
+  referring to S3 classes (#134). S4 unions are converted to R7 unions (#150).
+* Base numeric, atomic, and vector "types" are now represented as class unions
+  (#147).
 * Different evaluation mechanism for method dispatch, and greater restrictions 
   on dispatch args (#141)
 * `x@.data` -> `r7_data()`; probably to be replaced by casting.